landing-pages-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
aiosqlite==0.19.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
email-validator==2.1.0
jinja2==3.1.2
aiofiles==23.2.1
